Aplicacion del patron observador para el caso de uso Registro Suscripcion
Se agrego la validacion que primero se debe pagar la suscripcion para crearla, ademas se agrego una extension al tipo enumerado SuscriptionCostType para obtener un valor del tipo de costo, y se aplico el patron observador para el caso de uso Registrar suscripcion.

+++ lib/domain/interfaces/pay_method_interface.dart
@@ -1,10 +1,8 @@
 ///PayMethod: Es una interfaz utilizada para implementar métodos de pago
 abstract class IPayMethod {
-
   ///Permite realizar un pago a traves de un método especificado.
   ///[amount] es el monto a pagar.
-  Future<bool> getPayMethod(int amount) async {
+  Future<bool> getPayMethod(double amount) async {
     return true; //TODO: Implementar
   }
-
 }
